requirements.txt
feather-format
numpy==1.14.5
keras==2.2.2
tensorflow==1.10.0
dash==1.17.0
gunicorn
nltk
pandas==0.25.2
beautifulsoup4
matplotlib==3.1.3
lxml